ENT-4601 Move code into `FlowLogic` and improve Java API
Move the `await` functions into `FlowLogic` instead of relying on kotlin
extension functions. This means that they can be called from Java is a nicer
looking way.

Although the original `await` functions would compile in Java, they would
fail at runtime due to issues with Kyro and Quasar (lambda's need to be
specified as serializable due to being in the suspended stack). To get around
this a new interface `SerializableBiFunction` was created that was a function
and also extended `Serializable`. Overrides of `await` and `awaitFuture` were
added that used this new interface.
  - Kotlin now uses the original version (taking a function)
  - Java uses the new version (taking in `SerializableBiFunction`)

Tests were added to check the Java API.